adding some comments for better code readability
