+++ src/passes/conditionalSplitter/export.ts
@@ -1,2 +1,3 @@
-export * from './conditionalFunctionaliser';
+export * from './utils/conditionalFunctionaliser';
+export * from './utils/splitter';
 export * from './conditionalSplitter';
